% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionals.R
\name{kde2dplot}
\alias{kde2dplot}
\title{produces 2d kernel density plots.}
\usage{
kde2dplot(x, y = NULL, h, h.relative = 1, col = "red", add = FALSE,
  drawlabels = FALSE, n = 100, shade = TRUE, breaks = c(0.1, 0.25,
  0.5, 0.75, 0.9), ...)
}
\arguments{
\item{x}{x coords}

\item{y}{y coords}

\item{h}{bandwidths in x,y direction}

\item{h.relative}{adjustment to bandwidths relative to normal reference model. (For non-smooth data, will want < 1, probably.}

\item{col}{colour to use}

\item{add}{if FALSE, plot a whole new plot. Otherwise, superimpose on to existing plot.}

\item{drawlabels}{A value of false will give no labels on the contours whereas true will.}

\item{n}{resolution of underlying grid}

\item{shade}{shade in the interiors of the contours?}

\item{breaks}{breakpoints to plot contours at. These correspond to quantiles of the data - e.g. at 0.9, the space contained in the contours corresponds to an estimated 10 of the population, with the highest probability.}

\item{...}{additional params for contour see code{\link[graphics]{contour}}}
}
\description{
Creates a 2d kernel density plot.
}
\details{
Creates a 2d kernel density plot
}
\keyword{internal}
