% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\docType{data}
\name{thresholds}
\alias{thresholds}
\title{Definition of alpha values and thresholds for plotting ellipse confidence intervals}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\usage{
thresholds
}
\description{
Thresholds for shading of plots
}
\details{
Matrix of alpha values. The column names should be character strings 
giving the maximum value taken by upDown.sd*degrees.sd for each column. The 
rownames should be character strings giving the maximum numeric value taken by
the duration for each row.
}
\keyword{internal}
